tests/dialog-enabled.html

<!DOCTYPE html>
<html>
<head>
<meta charset=utf-8>
<script src="./dist/polyfill.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<dialog></dialog>
<script>
test(function() {
  dialog = document.querySelector('dialog')
  assert_true(dialog instanceof HTMLDialogElement);
}, "The DIALOG element should be recognized");
</script>
</body>
</html>
